Minor fixes in colors
